add boll controller
Ball starts it's movement at the start of the game and I generate a random boolean variable to decide which side the ball move first. According to this boolean a random vector is getting created and normalized in order to set the ball velocity constant.

I decided to change the vector direction in collision detection and not let Unity handles itself. Because honestly I couldn't find a proper way to prevent ball to lose its velocity.