Add paddle controller
Paddles are now following the ball in the Z axis. It first checks whether the ball coming towards the paddle or not and then adding force to the direction according to the Z axis position. If the ball is moving away from the paddle it tries to go back to the initial position on the Z axis.

Currently it is not working perfectly but improvments will be done after other parts has been developed.